requirements.txt
moviepy
openai-whisper
fpdf